% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ref_to_df.R
\name{build_data_ref}
\alias{build_data_ref}
\title{Build Data Reference Frame}
\usage{
build_data_ref(data, input = input)
}
\arguments{
\item{data}{The raw data from \code{\link[=find_ref]{find_ref()}} or \code{\link[=parse_ref]{parse_ref()}}.}
}
\value{
A tidy data frame with one row per reference and additional details.
}
\description{
This function converts the results from \code{\link[=find_ref]{find_ref()}} or \code{\link[=parse_ref]{parse_ref()}} into a tidy data frame.
It handles JSON structures and unnests them where possible to create a comprehensive reference data frame.
}
\details{
Internally used by \code{\link[=find_ref_to_df]{find_ref_to_df()}} and \code{\link[=parse_ref_to_df]{parse_ref_to_df()}}, this function processes the raw JSON or text data
into a structured data frame. It uses \code{safely_unnest} to handle nested lists within the data.
}
\keyword{internal}
