% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ref.R
\name{clean_ref}
\alias{clean_ref}
\title{Clean and Organize Reference Data}
\usage{
clean_ref(data)
}
\arguments{
\item{data}{A data frame of references to be cleaned, as produced by \code{\link[=find_ref_to_df]{find_ref_to_df()}} or \code{\link[=parse_ref_to_df]{parse_ref_to_df()}}.}
}
\value{
A cleaned and reorganized data frame with one row per reference and additional details.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function tidies up the data frame of references obtained from \code{\link[=find_ref_to_df]{find_ref_to_df()}} or \code{\link[=parse_ref_to_df]{parse_ref_to_df()}}.
It performs various cleaning and reorganization tasks to present the data in a more structured and readable format.

The function performs the following steps:
\itemize{
\item Collapses non-essential columns (i.e. other than \code{date} and \code{title}), combining multiple
entries into a single text string.
\item Reorganizes \code{author} information, creating a full name from given, particle, and family components. Also
add an \code{author_order} column. The column will remain in a list format.
\item Separates and cleans dates and titles, organizing them into primary and additional information.
The first date and first title are save in \code{date} and \code{title}. The other dates and titles are
saved in \code{other_date} and \code{other_title}.
\item Extracts the year from the date and places it in a separate column.
\item Removes extraneous punctuation and trims whitespace from character columns.
\item Relocates key columns to a more standardized order for easier analysis and readability.
}

This function can be used directly, or indirectly through \code{\link[=find_ref_to_df]{find_ref_to_df()}} or \code{\link[=parse_ref_to_df]{parse_ref_to_df()}}
by setting the \code{clean_ref} parameter to TRUE.
}
