% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ref.R
\name{collapse_non_essential_columns}
\alias{collapse_non_essential_columns}
\title{Collapse Non-Essential Columns (Internal)}
\usage{
collapse_non_essential_columns(data)
}
\arguments{
\item{data}{A data frame of references with potential list columns.}
}
\value{
A data frame with non-essential information collapsed.
}
\description{
Collapses non-essential columns in the reference data frame, converting lists and multiple entries into single text strings.
This is used as a preliminary step in cleaning and organizing the data.
}
\keyword{internal}
