% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ref.R
\name{find_ref}
\alias{find_ref}
\title{Find Bibliographic References in Documents Using Anystyle}
\usage{
find_ref(
  input = NULL,
  path = NULL,
  output_format = c("ref", "xml", "bib", "json"),
  no_layout = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{input}{Vector of file paths to the PDF documents to be analyzed.}

\item{path}{The path where the parsed file(s) will be saved. If NULL, files are saved
in the current working directory. If empty, the parsed data is returned directly and not saved.}

\item{output_format}{Desired output format, one of "ref", "xml", "bib", or "json". "ref" is a text file
with one reference per line. Default is "json".}

\item{no_layout}{Logical; if TRUE, anystyle '--no-layout' option is used (e.g., use this
if your document uses a multi-column layout). Default is FALSE.}

\item{overwrite}{Logical; if TRUE, existing files at the output location will be overwritten. Default is FALSE.}
}
\value{
If path is empty, returns a list (or a single element if only one input) containing the extracted references in the specified format. If path is specified, parsed files are saved at the location and the function returns NULL.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function utilizes the \href{https://anystyle.io/}{anystyle} Ruby gem to analyze PDF or text documents and
extract all references it finds. The input can be a single or multiple PDF documents,
}
\examples{
\dontrun{
# Find references in a single PDF and return as JSON
find_ref(
  input = "path/to/document.pdf",
  output_format = "json"
)

# Find references from a folder of documents and save as XML
find_ref(
  input = "path/to/documents/", path = "path/to/output",
  output_format = "xml"
)
}

}
