% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_anystyle.R
\name{install_anystyle}
\alias{install_anystyle}
\title{Install anystyle and anystyle-cli}
\usage{
install_anystyle()
}
\value{
Invisible NULL. The function is called for its side effects (installing software) rather than a return value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
This function installs 'anystyle' and 'anystyle-cli' using the system's Ruby gem package manager.
It requires Ruby and its gem package manager to be installed on the system.
\href{https://anystyle.io/}{anystyle} is a fast and flexible tool for parsing bibliographic references.

The function executes the 'gem install' command for 'anystyle' and 'anystyle-cli'.
It requires that the user has the necessary permissions to install gems on the system.
If Ruby or gem is not installed, the function will fail, and you'll need to install Ruby first.

Ensure that the Ruby environment is properly configured and that gem can be called from the command line.
See more details on Ruby's installation \href{https://www.ruby-lang.org/fr/downloads/}{here} and on
RubyGems \href{https://rubygems.org/pages/download?locale=fr}{here}.
}
\examples{
\dontrun{
install_anystyle()
}

}
