% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ranystyle-package.R
\docType{package}
\name{Ranystyle-package}
\alias{Ranystyle}
\alias{Ranystyle-package}
\title{Ranystyle: Automated Bibliographic Reference Parsing and Cleaning}
\description{
Provides tools for automated extraction, parsing, and cleaning of bibliographic references from PDF and text documents. Utilizing the 'anystyle' Ruby gem, it segments references and converts them into structured formats. The package offers functions to reorganize and clean the data, making it easier to analyze and use bibliographic information. It handles various input and output formats and provides utilities for detailed reference management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agoutsmedt/Ranystyle}
  \item Report bugs at \url{https://github.com/agoutsmedt/Ranystyle/issues}
}

}
\author{
\strong{Maintainer}: Aurélien Goutsmedt \email{aurelien.goutsmedt@uclouvain.be} (\href{https://orcid.org/0000-0002-3788-7237}{ORCID})

}
\keyword{internal}
