% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ref.R
\name{reorganize_author}
\alias{reorganize_author}
\title{Reorganize Author Information (Internal)}
\usage{
reorganize_author(data)
}
\arguments{
\item{data}{A data frame of references with an 'author' column.}
}
\value{
A data frame with reorganized author information.
}
\description{
Reorganizes the 'author' column in the reference data frame, combining given, particle, and family name components into full names.
This helps standardize author information for further analysis and display.
}
\keyword{internal}
