% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ref_to_df.R
\name{safely_unnest}
\alias{safely_unnest}
\title{Safely Unnest Data}
\usage{
safely_unnest(data)
}
\arguments{
\item{data}{A data frame potentially containing list columns.}
}
\value{
A data frame with nested data unnested where it was safe to do so.
}
\description{
This helper function is used to unnest data frames that contain list columns, doing so only when it's safe (i.e., the list has no more than one element per row).
}
\keyword{internal}
