% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_ref.R
\name{separate_data_and_title}
\alias{separate_data_and_title}
\title{Separate and Clean Dates and Titles (Internal)}
\usage{
separate_data_and_title(data)
}
\arguments{
\item{data}{A data frame of references with 'date' and 'title' columns.}
}
\value{
A data frame with separated and cleaned date and title information.
}
\description{
Separates primary and additional date and title information in the reference data frame and performs basic cleaning operations.
This helps clarify the key information for each reference and removes unnecessary characters.
}
\keyword{internal}
